feat: Phase 112F __main__.py + remaining type error fixes (165→0 errors)
Fixed all 165 type errors in katrain/__main__.py and 6 errors in other files:

**katrain/__main__.py** (165 errors → 0):
1. **Import cleanup**: Changed `from typing import List, Optional, Set` to `from typing import Any`
2. **KaTrainGui.__init__**: Added `**kwargs: Any) -> None` annotation
3. **Instance variables**:
   - `engine: Any = None`
   - `message_queue: Queue[Any] = Queue()`
   - `_clock_events: list[Any] = []`
4. **Config methods**: Added return types and parameter types:
   - `_load_export_settings() -> dict[str, Any]`
   - `_save_export_settings(sgf_directory: str | None, selected_players: list[Any] | None) -> None`
   - `_save_batch_options(options: dict[str, Any]) -> None`
5. **State subscription methods**: Added type annotations:
   - `_setup_state_subscriptions() -> None`
   - `_on_game_changed(event: Any) -> None`
   - `_on_analysis_complete(event: Any) -> None`
   - `_on_config_updated(event: Any) -> None`
   - `_do_ui_update(dt: Any) -> None`
6. **Popup factory methods**: Added `-> Any` return types and parameter types
   - `_create_new_game_popup()`, `_create_timer_popup()`, etc.
   - Added `popup: Any` type annotation for getattr results
7. **Log method**: Fixed signature `log(message: str, level: int = OUTPUT_INFO) -> None`
8. **type: ignore annotations**: Added for manager callbacks and Game/GameProtocol incompatibility
9. **All _do_* command methods**: Added comprehensive type annotations (80+ methods)
10. **KaTrainApp methods**: Added type annotations to build(), on_language(), etc.

**katrain/tools/batch_analyze_sgf.py** (2 errors → 0):
- Added `-> None` to `main()` function
- Added `-> None` to nested `log_print()` function

**katrain/gui/error_handler.py** (1 error → 0):
- Changed `log_level: float` to `log_level: int` in both `handle()` and `_handle_impl()`

**katrain/gui/features/commands/popup_commands.py** (3 errors → 0):
- Added `assert ctx.config_popup is not None` to narrow type before attribute access

**Progress**:
- Phase 112F: 165 + 6 = 171 errors → 0 errors (100% fixed)
- Total project: 1352 → 0 errors (100% complete)
- mypy check: Success: no issues found in 205 source files

All 3776 tests pass.

+++ katrain/gui/error_handler.py
@@ -39,7 +39,7 @@ class ErrorHandler:
         error: Exception,
         *,
         notify_user: bool = True,
-        log_level: float = OUTPUT_ERROR,
+        log_level: int = OUTPUT_ERROR,
         fallback_message: str = "An error occurred",
     ) -> None:
         """Handle an exception with logging and optional user notification.
@@ -62,7 +62,7 @@ class ErrorHandler:
         self,
         error: Exception,
         notify_user: bool,
-        log_level: float,
+        log_level: int,
         fallback_message: str,
     ) -> None:
         """Internal implementation of handle()."""
